*Issue #, if available:*
#2756 

*Description of changes:*

1. persistent_workers=True if num_workers > 0.
2. reload_dataloaders_every_n_epochs=0.
3. Remove the additional evaluation in top_k_average. Use the best validation score from training instead to reduce overhead.

This PR can substantially reduce the time overhead of multi-gpu training introduced by 0.6.2. Specifically, for 60s time limit, it even uses less training time compared to 0.5.2.

**Environment**
g4dn.12xlarge, pytorch-lightning 1.8.6

**time_limit=30**

* best
    * PR's time: 76.5s
    * 0.5.2 time: 75.6s
* greedy soup
    * PR's time: 87.3s
    * 0.5.2 time: 82.2s
* uniform soup
    * PR's time: 77.1s
    * 0.5.2 time: 75.9s

**time_limit=60**

* best
    * PR's time: 104.7s
    * 0.5.2 time: 110.8s
* greedy soup
    * PR's time: 116.7s
    * 0.5.2 time: 120.0s
* uniform soup
    * PR's time: 105.9s
    * 0.5.2 time: 111.7s

Here is the code to reproduce results:
```
train_dataset, _, test_dataset = ImageDataset.from_folders(
        "https://autogluon.s3.amazonaws.com/datasets/shopee-iet.zip")

    version = core.version.__version__
    predictor = MultiModalPredictor(label="label", path=f"./automm_imgcls/{version}/{time.time()}")
    print("starting fit\n")
    start = time.time()
    predictor.fit(
        train_data=train_dataset,
        hyperparameters={
            "optimization.top_k_average_method": "best",  # "uniform_soup" or "greedy_soup"
        },
        time_limit=30,  # or 60
    )  # you can trust the default config, e.g., we use a `swin_base_patch4_window7_224` model
    end = time.time()
    print(f"version {version} takes {end - start} seconds")
```

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.